freeze interns2 mtp delivery code

+++ lmdeploy/pytorch/models/qwen3_moe.py
@@ -212,7 +212,7 @@ class Qwen3MoeSparseMoeBlock(nn.Module):
             self.hidden_dim,
             self.num_experts,
             bias=False,
-            dtype=dtype,
+            dtype=torch.float32,
             device=device,
             is_tp=False,
         )
